% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation Test For Proportions}
\usage{
permutation_test(
  sc_utils_obj,
  sample_1 = NA,
  sample_2 = NA,
  sample_identity = "orig.ident",
  n_permutations = 1000
)
}
\arguments{
\item{sc_utils_obj}{sc_utils object}

\item{sample_1}{First sample to compare (ie. control)}

\item{sample_2}{Sample to compare to first sample (ie. treatment)}

\item{sample_identity}{Column that has sample names}

\item{n_permutations}{Number of permutations}

\item{cluster_identity}{Column that has cluster names}
}
\description{
Permutation Test For Proportions
}
