% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_utils.R
\docType{class}
\name{sc_utils-class}
\alias{sc_utils-class}
\alias{sc_utils}
\title{Single-Cell Utilities}
\usage{
sc_utils(seurat_object)
}
\arguments{
\item{seurat_object}{Seurat object with meta-data}
}
\description{
Single-Cell Utilities

Single-Cell Utilities Constructor
}
\section{Slots}{

\describe{
\item{\code{meta_data}}{Seurat meta-data}

\item{\code{results}}{Results for various analysis}
}}

