% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plot.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Plot Permutation Results}
\usage{
permutation_plot(
  sc_utils_object,
  FDR_threshold = 0.05,
  log2FD_threshold = log2(1.5),
  order_clusters = TRUE
)
}
\arguments{
\item{FDR_threshold}{FDR value cutoff for significance}

\item{log2FD_threshold}{Absolute value of log2FD cutoff for significance}

\item{order_clusters}{Whether to order the clusters by observed log2FD}

\item{sc_utils_obj}{sc_utils object}
}
\description{
Plot Permutation Results
}
